git-svn-id: https://svn.alfresco.com/repos/alfresco-enterprise/alfresco/HEAD/root@10709 c4b6b30b-aa2e-2d43-bbcb-ca4b014f7261

+++ source/java/org/alfresco/repo/domain/hibernate/OldADMPermissionsDaoComponentImpl.java
@@ -93,7 +93,7 @@ public class OldADMPermissionsDaoComponentImpl extends AbstractPermissionsDaoCom
         if (acl != null)
         {
             // maintain referencial integrity
-            getACLDAO(nodeRef).setAccessControlList(nodeRef, null);
+            getACLDAO(nodeRef).setAccessControlList(nodeRef, (Long) null);
             aclDaoComponent.deleteAccessControlList(acl.getId());
         }
     }
